Moved cogt tests under pipelex
