+++ tests/pipelex/cogt/asynch/imgg/test_imgg_openai.py
@@ -5,8 +5,8 @@ from pipelex.cogt.llm.llm_models.llm_platform import LLMPlatform
 from pipelex.cogt.plugin.openai.openai_factory import OpenAIFactory
 from pipelex.tools.misc.base_64_utils import save_base64_to_binary_file
 from pipelex.tools.misc.file_utils import ensure_path, get_incremental_file_path
-from tests.cogt.test_data import IMGGTestCases
 from tests.conftest import TEST_OUTPUTS_DIR
+from tests.pipelex.cogt.test_data import IMGGTestCases
 
 
 @pytest.mark.imgg
