+++ tests/pipelex/cogt/asynch/llm_async/test_llm_engines.py
@@ -11,7 +11,7 @@ from pipelex.cogt.llm.llm_models.llm_platform import LLMPlatform
 from pipelex.cogt.llm.llm_worker_abstract import LLMWorkerAbstract
 from pipelex.cogt.llm.llm_worker_factory import LLMWorkerFactory
 from pipelex.hub import get_inference_manager, get_llm_deck, get_llm_models_provider, get_report_delegate
-from tests.cogt.test_data import LLMTestConstants, Person
+from tests.pipelex.cogt.test_data import LLMTestConstants, Person
 
 
 @pytest.mark.llm
