+++ tests/pipelex/cogt/asynch/llm_async/test_llm_gen_object.py
@@ -8,7 +8,7 @@ from pipelex import pretty_print
 from pipelex.cogt.llm.llm_job_components import LLMJobParams
 from pipelex.cogt.llm.llm_job_factory import LLMJobFactory
 from pipelex.hub import get_llm_deck, get_llm_worker
-from tests.cogt.test_data import LLMTestCases
+from tests.pipelex.cogt.test_data import LLMTestCases
 
 
 def get_async_worker_and_job(llm_preset_id: str, user_text: str):
