+++ tests/pipelex/cogt/asynch/llm_async/test_llm_vision.py
@@ -10,7 +10,7 @@ from pipelex.cogt.llm.llm_job import LLMJobParams
 from pipelex.cogt.llm.llm_job_factory import LLMJobFactory
 from pipelex.hub import get_llm_worker
 from pipelex.tools.misc.base_64_utils import load_binary_as_base64
-from tests.cogt.test_data import LLMVisionTestCases
+from tests.pipelex.cogt.test_data import LLMVisionTestCases
 
 
 @pytest.mark.llm
