+++ tests/pipelex/cogt/asynch/ocr/test_ocr.py
