+++ tests/pipelex/cogt/conftest.py
@@ -168,20 +168,6 @@ def llm_id(request: pytest.FixtureRequest) -> str:
 
 @pytest.fixture(
     params=[
-        # ImggHandle.FLUX_1_PRO_LEGACY,
-        # ImggHandle.FLUX_1_1_PRO,
-        # ImggHandle.FLUX_1_1_ULTRA,
-        ImggHandle.SDXL_LIGHTNING,
-        # ImggHandle.OPENAI_GPT_IMAGE_1,
-    ]
-)
-def imgg_handle(request: pytest.FixtureRequest) -> ImggHandle:
-    assert isinstance(request.param, ImggHandle)
-    return request.param
-
-
-@pytest.fixture(
-    params=[
         LLMJobParams(
             temperature=0.5,
             max_tokens=None,
