+++ tests/pipelex/cogt/synchro/llm_providers/test_mistral.py
