+++ tests/pipelex/cogt/synchro/serde/conftest.py
