+++ tests/pipelex/cogt/synchro/serde/test_serde_llm_classes.py
