+++ tests/pipelex/cogt/synchro/test_llm_model_provider.py
