+++ tests/pipelex/cogt/synchro/test_setup_workers.py
